refactor: 책임 분리
ConsoleInputview가 너무 많은 책임을 지니고 있어,
ConsoleInputView가 지니고 있던 유효성 검사 책임을
InputItemInfos에 할당
